feat(ui): Компонент Image
