+++ packages/ui/src/components/Image/index.ts
@@ -1,0 +1,2 @@
+export { Image } from './Image';
+export type { ImageProps } from './Image';
